requirements.txt
click==8.0.4
dill==0.3.4
fastparquet==0.8.1
GitPython==3.1.27
ipykernel==6.9.1 
matplotlib==3.5.2
pandas==1.4.2
lightgbm==3.2.1
pyarrow==8.0.0
PyYAML==6.0
scikit-learn==1.1.1
scikit-optimize==0.9.0
statsmodels==0.12.2
tqdm==4.64
unidecode==1.3.4
xgboost==1.6.1